backend/requirements.txt
Django>=5.1.4
djangorestframework>=3.15.2
django-cors-headers>=4.6.0
python-dotenv>=1.0.0
Pillow>=10.1.0  # For image processing
gunicorn>=21.2.0  # For production deployment
